fix test and move generators to a new module
