+++ helios-core/src/main/kotlin/helios/typeclasses/Decoder.kt
@@ -1,6 +1,7 @@
 package helios.typeclasses
 
 import arrow.core.Either
+import arrow.higherkind
 import helios.core.Json
 
 sealed class DecodingError
@@ -9,6 +10,7 @@ data class BooleanDecodingError(val value: Json) : DecodingError()
 data class NumberDecodingError(val value: Json) : DecodingError()
 data class KeyNotFound(val name: String) : DecodingError()
 
+@higherkind
 interface Decoder<out A> {
   fun decode(value: Json): Either<DecodingError, A>
 }
