+++ helios-core/src/main/kotlin/helios/typeclasses/Encoder.kt
@@ -5,6 +5,5 @@ import helios.core.Json
 
 interface Encoder<in A> {
   fun A.encode(): Json
-  fun
-      Iterable<A>.toJson(): JsArray = JsArray(map { it.encode() })
+  fun Collection<A>.toJson(): JsArray = JsArray(map { it.encode() })
 }
