+++ helios-optics/src/test/kotlin/helios/optics/domain.kt
@@ -1,6 +1,7 @@
 package helios.optics
 
 import helios.meta.json
+import helios.test.generators.alphaStr
 import io.kotlintest.properties.Gen
 
 @json
@@ -13,6 +14,6 @@ data class Street(val name: String) {
   companion object
 }
 
-fun genStreet(): Gen<Street> = Gen.string().map { str -> Street(str) }
+fun genStreet(): Gen<Street> = Gen.alphaStr().map { str -> Street(str) }
 
 fun genCity(): Gen<City> = Gen.list(genStreet()).map { streets -> City(streets) }
