+++ helios-optics/src/test/kotlin/helios/optics/instances.kt
@@ -2,7 +2,10 @@ package helios.optics
 
 import arrow.core.None
 import arrow.core.Option
+import arrow.core.extensions.option.eq.eq
 import arrow.core.orElse
+import helios.test.generators.alphaStr
+import helios.test.generators.*
 import arrow.test.UnitSpec
 import arrow.test.generators.functionAToB
 import arrow.test.generators.option
@@ -14,6 +17,8 @@ import arrow.typeclasses.Monoid
 import helios.core.JsArray
 import helios.core.JsObject
 import helios.core.Json
+import helios.instances.jsobject.eq.eq
+import helios.instances.json.eq.eq
 import helios.optics.jsarray.each.each
 import helios.optics.jsarray.index.index
 import helios.optics.jsobject.at.at
@@ -27,7 +32,7 @@ class InstancesTest : UnitSpec() {
 
     testLaws(
       OptionalLaws.laws(
-        optionalGen = Gen.string().map { JsObject.index().index(it) },
+        optionalGen = Gen.alphaStr().map { JsObject.index().index(it) },
         aGen = Gen.jsObject(),
         bGen = Gen.json(),
         funcGen = Gen.functionAToB(Gen.json()),
@@ -72,16 +77,17 @@ class InstancesTest : UnitSpec() {
     )
 
     testLaws(LensLaws.laws(
-      lensGen = Gen.string().map { JsObject.at().at(it) },
+      lensGen = Gen.alphaStr().map { JsObject.at().at(it) },
       aGen = Gen.jsObject(),
       bGen = Gen.option(Gen.json()),
       funcGen = Gen.functionAToB(Gen.option(Gen.json())),
-      EQA = Eq.any(),
-      EQB = Eq.any(),
+      EQA = JsObject.eq(),
+      EQB = Option.eq(Json.eq()),
       MB = object : Monoid<Option<Json>> {
         override fun Option<Json>.combine(b: Option<Json>) = orElse { b }
         override fun empty(): Option<Json> = None
       }
     ))
+
   }
 }
