+++ helios-test/build.gradle
@@ -1,0 +1,10 @@
+
+dependencies {
+    compile "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlinVersion"
+    compile "io.arrow-kt:arrow-test:$arrowVersion"
+
+    compile project(":helios-core")
+}
+
+apply from: rootProject.file('gradle/gradle-mvn-push.gradle')
+apply from: rootProject.file('gradle/generated-kotlin-sources.gradle')
