+++ helios-test/gradle.properties
@@ -1,0 +1,4 @@
+# Maven publishing configuration
+POM_NAME=Helios Test
+POM_ARTIFACT_ID=helios-test
+POM_PACKAGING=jar
