+++ settings.gradle
@@ -9,3 +9,4 @@ include ':helios-sample'
 include ':helios-optics'
 include ':helios-dsl-syntax-compiler'
 include 'helios-docs'
+include 'helios-test'
